Use manually require spec_helper.js in specs
... rather than having mocha `--require spec/javascript/setup.js` from
the command to run tests. This will allow us to use `beforeEach` etc in
`spec_helper.js` to add global setup hooks. In this case, I've added a
beforeEach hook that will fail specs if anything is logged with
`console.error` (such as the `TypeError: Cannot read property 'needed'
of undefined` error that is being logged for my one current test, and
which is a significant problem with the spec, but which previously
wasn't failing the spec).