Allow drag-and-drop of Grocery items from one store to another
Uses (more-or-less) the native HTML5 drag and drop API, via view event
handlers and then passing the update via Vuex.

For the moment, I think that this probably was the right decision vs
using some plugin library. I don't think I needed a lot more than the
native drag-and-drop API, plus I got to learn it (or, at least, part of
it), which I haven't used before.

Tests are still missing, though. But that's because I'm trying to ship
this before I go to the store! :-p :)