Use only built-in fonts in home splash view
Previously, some of the fonts in the home splash view were using a
webfont (Karla, sourced via Google), and others were using a built-in
font (monospace). As a result, the monospace text was rendering sooner,
because the Karla text had to wait for the webfont to load before it
could render. This was an unpleasing visual effect.

This change should make it so that:
1. All text in the splash view renders at the same time
2. The perceived load time will be faster