Pause before rendering Vue app in development to prevent FOUC
Previously, we were experiencing a FOUC (flash of unstyled content) in
development because it was taking Chrome a while to parse the CSS source
maps that were recently added in f5efc5c / #35. At least, the FOUC seems
to be fixed if I disable CSS source maps, so I'm _assuming_ that's the
problem.

To facilitate this, I've extracted a `renderApp` function into our
`app/javascript/vendor/customized_vue.js`, which is a nice DRYing of the
code. :)