Remove references to application.js/common.js in test config
`application.js` was deleted in #46 / bca1408 .

`shared/common.js` is now loaded in
`app/javascript/vendor/customized_vue.js`, so there's no longer a need
to manually load it in tests as the entry point of the test webpack
config.

I didn't notice either of these necessary/good changes in time to
include them in #48 / 1b5fc15 because that was a hotfix for a production
error that I wanted to get out right away.