Render modal DOM elements at end of Store, not next to button
This fixes a visual bug: http://take.ms/RV7P4

The problem is that previously, when the modal opened, the modal DOM
content was being insterted between two `el-button`s. This was then
breaking (by making it no longer apply) a `.el-button + .el-button`
style rule that Element UI was thoughtful enough to provide for me. :)

By inserting the modal DOM content at the end of the Store instead of
next to the button that triggers the modal to open, this CSS rule will
be preserved even when the modal is open.