Delete stores via the Vuex deleteStore action
And add some tests for once.

The lesson that I learned is that there doesn't seem to be a
straightforward way, when testing a Vue component, to make assertions
that a given mutation was called. Instead, one should assert on the
effects of those actions, I guess. I can live with that, since in some
ways it's more integration test-y rather than unit test-y, and
integration tests often seem to provide better coverage against real
bugs, and also less test brittleness (since they are less connected to
implementation details).