Add Foreman to Gemfile
The app failed to boot when we deployed #162 with error:
```
Dec 16 23:28:59 davidrunger app/web.1: bash: foreman: command not found
```

I had avoided adding `foreman` to the Gemfile, since the Foreman docs
say explicitly not to add it. However, hopefully `require: false` makes
this not a problem?

I figured that Heroku might make the `foreman` command available, since
I had thought that's what it uses to run the `Procfile`, but apparently
one or both of these things is not true.