Load all <script> tags with defer
In practical terms, I don't think that this will atually matter hardly
at all (and this seems to be confirmed by my testing with Chrome dev
tools w/ Network throttled to "Fast 3G" and CPU throttled to "6x
slowdown"). Loading/parsing JavaScript with `defer` basically means, as
far as I can tell, that HTML parsing is blocked for as long as it would
be otherwise. However, since everything at davidrunger.com is basically
rendered via JavaScript, there is hardly any HTML to parse in the first
place. But who knows, maybe there will be in the future (e.g. if I setup
prerendering or server-side rendering). And it just seems like a good
practice as a general rule. So I'm doing it.

Reference: http://www.growingwiththeweb.com/2014/02/async-vs-defer-attributes.html