Install administrate from RubyGems
There is a breaking change on administrate `master` (see [revert PR][1])
that appears not to be on RubyGems. This should be in general more
stable going forward, anyway.

[1]: https://github.com/davidrunger/david_runger/pull/184

The reason that I had been on `master` originally is because there
wasn't a gem release that supported Rails 5.1, but now there is, so
there's no longer a good reason to stay on master.