Modify migration to not enable pg_stat_statements on Heroku
I know it's not a good idea to modify migrations retroactively, but this
migration can't run successfully as-is on Heroku (and is thus blocking
deploys) because the `pg_stat_statements` extension is already enabled
on Heroku. Adding `IF NOT EXISTS` should prevent this migration from
causing an error when run on Heroku.