Re-enable Vue devtools in development
I recently set `Vue.config.devtools = false` because doing so suppressed
a warning/notice about devtools in the output when running my JavaScript
tests. I thought that I was just disabling that notice, but it turns out
that setting that to `false` also actually turns of the devtools
functionality. So this re-enables the devtools in `development`.