Use `style` loader, not new webpacker `css`/`sass` loaders
I initially only updated the development environment with this change
from `style` => `css` & `sass` ... but thinking about it more, I think
it makes sense to just have a single `style` loader, as it was
originally. Therefore, this PR reverts the change to development and
nullifies the webpacker `css` and `sass` loaders, so I will continue
using the `style` loader in all environments.