Fix element-ui CSS
Man, keeping this CSS configuration working is tough.

I think that the main change in this PR is adding the `allChunks: true`
option to the `ExtractTextPlugin` configuration.

Per the docs:

> `allChunks`: Extract from all additional chunks too (by default it
> extracts only from the initial chunk(s))

> When using `CommonsChunkPlugin` and there are extracted chunks (from
> `ExtractTextPlugin.extract`) in the commons chunk, `allChunks`
> **must** be set to `true`

This was working previously without the `allChunks` option, but
something (or some combination of things) in the package upgrades that
I've done recently seems to have broken it. But now with this PR it's
working again. (The "brokenness" is that actually 2 `commons-<hash>.css`
bundles were being built on production. One contained the element-ui
CSS; the other didn't -- and the one that did _not_ contain the element-
ui CSS is the one that was being served on production.)

This _might_ (total speculation) solve the problem of the fact that _I
think_ (at least, according to dareboost.com) that I might have been
loading the element-ui CSS rules twice?