Don't load element-ui button CSS twice
It turns out that the element-ui `Button` element CSS seems to be
included in the `dropdown.css`. I'm not sure if this is intentional or
not, but anyway it has the effect that the `button` CSS gets loaded
twice if imported once via `Button`/`button.css` and then a second time
through `Dropdown`/`dropdown.css`. So this PR switches from using the
`babel-plugin-component` plugin to instead explicitly import the CSS
that I want (thus giving me the option to _not_ import the `button.css`,
since it's also included in `dropdown.css`).