Remove file option from vue-loader config that does nothing
I'm not sure if this used to do anything before, but it certainly
doesn't seem to be doing anything now (nor is it mentioned in the [vue-
loader docs][1]... although nor is `sass` or `scss` mentioned there, but
those options _do_ seem to have an effect, although I believe that's
because they are the `lang` attribute of many/all of my `<style>` tags,
which cannot be said of `file`). To prove this, I replaced `file-loader`
with `runger- loader` and `bin/webpack-dev-server` compiled just fine.
Also, images still loaded in the app.

[1]: https://vue-loader.vuejs.org/en/options.html#loaders