Use when-dom-ready lib to check for DOM readiness
Previously, I was checking for the `DOMContentLoaded` event, but this is
not ideal because if that event has _already_ fired, then the callback
will not be triggered, whereas with `when-dom-ready`, [it's claimed][1]
that

> The Promise will resolve instantly if the DOM is already ready.

[1]: https://github.com/lukechilds/when-dom-ready#when-dom-ready

(Looking at the code, I _believe_ that this will apply to the "callback"
way to use `when-dom-ready`, in addition to the "Promise-based" usage
that is mentioned in the docs.)

Thanks, `when-dom-ready`! :smile: :+1:

An additional part of my motivation is that I'm also going to need a
DOM-ready listener function for another upcoming PR (in order to lazy-
load below-the-fold images only after the DOM has loaded / become
interactive).