Use WebP image (if supported) for Groceries app background image
Also, re-encode the image as WebP with a quality of 85 rather than what
I believe had been done previously of 92. These 7 points of difference
significantly change the size of the WebP file without making a
significant negative impact on the user experience, particularly since
this image is just a background image, anyway.