Lazy-load images on Home page
This is a recommendation that comes from Chrome's Lighthouse "Audit" in
the Chrome Developer Tools.

I'm not sure what exactly is the time at which the images are considered
to be sufficiently lazily loaded to meet the Chrome audit, but it seems
in my case that a 3000ms timeout is required in order to satisfy the
audit. It's somewhat unfortunate to add another `setTimeout` to the
codebase, but what can ya do.