Send release events to graphite
This way, I will be able to see releases in my Grafana graphs.

Also, move the `db:migrate` rake command from a deploy command in
`.travis.yml` to the new `bin/release-tasks` script.

Also, change `bin/_webpack-setup.rb` to executable, just because
everything else in `bin/` is executable, and OCD. If it doesn't need to
be executable, it probably shouldn't be ... but, like I say, OCD. That's
the fun of this just being a personal project and it not actually
mattering. :)