Move webpack-cli from devDependencies to dependencies
We need it on Heroku to compile assets. Currently asset compilation
(and, by extension, deploys) are failing on Heroku/production.

It's not totally my fault. At least, _I think_ that wherever I read
about `webpack-cli` the suggestion was to add the `-D` flag to `yarn
add`. I can't be blamed for mindlessly following instructions from the
Internet, can I? :-p