Rename Clipboard to ClipboardJS
Per the [clipboard.js 2.0.0 release][1], the documented/advised naming
convention is to use `ClipboardJS` to avoid conflict with the "new"
(though I haven't actually looked how "new" it is) `window.Clipboard`
native function.

[1]: https://github.com/zenorocha/clipboard.js/releases/tag/v2.0.0

This name change doesn't actually affect anything / my usage of
clipboard.js was working fine even without this change (since the import
was getting transpiled to a variable name other than `Clipboard`,
anyway) but it seems like a good idea to be consistent with the docs and
to not create confusion.