Make Request::recent scope ignore daylight savings time change
When I ask for `Request.recent(2.days)`, I think that I really want
`Request`s created within the last 48 hours, not just since the same
time of day two calendar days ago. This change updates the application
code to reflect that desire.

The real motivation for getting involved with this code/specs is that
there is currently a failure on `master` related to daylight savings
time; this change fixes that failure.