Install statsd from davidrunger fork
My [`davidrunger/statsd` fork][1] drops the optional `winser` dependency
of `statsd`. Motivation: `winser` depends on `sequence` which references
the `ender` engine which raises a warning in `yarn check`.

`modern-syslog` was also causing a warning of `warning Optional
dependency "statsd#modern-syslog" not installed` (which is weird,
because it was still listed in my `yarn.lock`, but whatever), so I've
dropped that, too.

[1]: https://github.com/davidrunger/statsd