Log errors via Rails.logger (in additionl to Rollbar)
Rollbar is not going to be 100% reliable, so it's good to have
`Rails.logger` logging available as a backup. Plus, logs are to capture
what happens in the app; errors are notable happenings. Also, it might
be easier in some cases to search logs than to search Rollbar. Finally,
I believe that I have indefinite logs retention on my DigitalOcean
droplet, whereas I don't think that Rollbar retains errors indefinitely.