Create a SkillRow Vue component, rather than writing raw HTML
I'm not totally sure that this is worthwhile / a net positive change. I
felt that there was a little too much repetition in the raw HTML that we
were previously using for these skill rows, though. By extracting a
`SkillRow` component, that component can DRYly do certain things for us,
like determining the devicon class name.