Specify webpack version with caret
(No particular reason, except for consistency with how our other
packages are specified, and to ensure that dependabot will update this
package, as it updates our other packages that are specified with
carets. I had originally introduced the non-caret specification, I
think, by saying `yarn add webpack@4.4.x`. Come to think of it... that
was because there was an issue with webpack 4.5 regarding some certain
thing (which I encountered while bumping `vue-loader` from 14 to 15).
Hopefully 4.6 fixes whatever that issue was...)