Remove newline from within pug template inline tag
Apparently this is not allowed / break compilation of the template.
Fortunately this problem occurred during deploy rather than on
production. However, this was not caught by tests. This PR also adds a
basic test for the Home Vue component that should catch this sort of
issue in CI in the future.