Rubocop: Disable linting of db/schema.rb
This file is auto-generated by Rails, and frequently does not conform to
our Rubocop rules. Since we are not really free to modify this auto-
generated file to conform to Rubocop (and since it would be inefficient
to do so, even if we could), let's disable Rubocop linting of this file.