Reload log app after entering any new data
This change makes it such that the user does not need to manually reload
the page in order to see any data that they have entered. This is a
temporary measure until I make an effort to add the new data directly to
the Vuex store at the same time that we POST the data to the server.