Add BriefTimeField field type to admin; use for requests.requested_at
This is to make more room for the upcoming "ISP" column in the index.

An alternative would be to use Administrate's
`Field::DateTime.with_options(format: '%b %-d %l:%M%P')`, but then we
don't get the full time when hovering over the time, which we _do_ get
using this implementation (by providing a `title` attribute for the
`<span>`s).