Store raw user agent in requests
A subsequent commit will use `browser` to parse the user agent live for
display in admin. This will make admin views of the request index and
show pages slower, but it will speed up actual user requests to the rest
of our pages, since we won't be doing user agent parsing blocking the
response to users.