Remove `requests.bot` column and usage in code
First, I don't even really use this column anymore; it's basically
redundant with the parsed user agent column that we are now showing on
the Request admin show & index pages.

Second, even if I were still using this, then it would probably be
better to compute this value live, rather than cacheing it in the
database as we have been doing up until now (this PR).