Add Log#log_entries_ordered association (for performance)
This allows us to eager-load a log's log-entries for serializing into
the bootstrap data.

The reason that these need to be ordered on the server side (as opposed
to just letting the client do it) is because we might want to send only
the last 3 entries, if it's a text log.

Also, eager-load the Log#log_inputs association.

This fixes (removes) the two N+1 queries that previous existed in the
logs#index action.