Serialize LogEntry created_at time using raw timestamp from DB
This results in a significant performance increase (when there are a lot
of log entries), since we skip parsing many DB timestamps into `Time`
objects, which is actually fairly slow, relatively speaking.

This change produces a slightly degraded user experience for users who
are in Pacific Time (the application's time zone), since times are now
rendered in UTC time rather than in Pacific Time. This might be a better
experience for users in non-Pacific time zones, though, for whom UTC
time might be more useful than Pacific Time. Anyway, the ultimate fix is
probably that the client should parse these times into the user's time
zone?

Also, fix ActiveRecord config to assume that database timestamps are in
UTC (as they are, at least on Heroku/production) rather than to assume
that the database timestamps are in local time. (This wasn't causing
bugs, since in Heroku the local machine time was in UTC, so `:local` and
`:utc` were effectively the same setting on production, but it was
causing issues after pulling the database down locally (with UTC
timestamps), since my local machine is set to Pacific Time.)