Remove JSX and JSX dependencies from the app
This was causing issues for some reason when trying to upgrade
webpacker. This seems to be surfacing some sort of incompatibility with
Vue and the cutting edge babel ecosystem. I dunno. Since I _barely_ use
JSX, it's not worth trying to keep. In the two instances where I was
using JSX, it was plenty easy enough to just write the raw `render`
function code. Maybe this will even make the app a little smaller?,
although probably not because I think the JSX was being compiled down to
the same thing, anyway, as part of the webpacker/babel compilation.