Request user's phone number when user asks for text without one
There is a serious need for refactoring/cleanup re:
`setShowModal`/`setShowNeedPhoneNumberModal`. Rather than having
separate attributes in the store for
`showModal`/`showNeedPhoneNumberModal`, these should probably be
consolidated into a single action like `openModal(modalName)` that
maniuplates a single data structure that contains a list of open modals
(and the order in which they were opened, in order to display them in
the correct "order", i.e. most-recently opened modal "on top").