Refactor modal display logic and data structure
This commit refactors the modal display logic and data structure,
generalizing it to accommodate an arbitrary number of modals (as long
as each modal is given a unique `name` prop).

This refactor also eliminates the necessity of adding a data attribute
to indicate whether each specific modal is showing / not showing and
eliminates the necessity of adding mutations for showing/hiding each
modal that might be used.

There will be a slight performance penalty, in that we are now doing
some array scanning (and other array operations) rather than
manipulating specific booleans indicating the show/hide state of each
individual modal, but this should be negligible. This performance
penalty could be minimized somewhat by storing the show/hide state of
each modal in an object rather than in an array, but then that would
somewhat complicate the `hideTopModal` mutation, in contrast to the
simplicity of just using an array as a stack where we know/assume that
the last element in the array is the most recently opened modal.