Fetch all log entries on page load
In a39b4e6 / #1486 , we stopped including log entries in the bootstrap
data (for a faster initial page load).

This has the downside, though, that the user might have to wait a little
while for the log entries to be loaded after viewing a log.

This PR adds an immediate fetch to eager-load the log entries for all
logs after page load. Thus, hopefully by the time that the user selects
a log, the log entries for that log (and all other logs) will have
already been loaded, and the user won't experience any delay in fetching
the log entries, and will solely experience a performance enhancement
from #1486 in that the log entries are no longer serialized prior to
page load.