Focus log entry input whenever log/route is changed
Previously, only the initial log that was viewed would have its input
autofocused (which was done via the `mounted` lifecycle hook). Since vue
router is now re-using the same component(s), the `mounted` hook was not
being called when viewing any subsequent logs, and thus the log entry
input was not being autofocused.