Refactor data model behind logs
This is a pretty big backend change. Hopefully there are no user-facing
changes! :)

Primarily the change here is to:
1. drop the concept of "log inputs" and instead assume that each log
   only has one input; the information about this input is now stored on
   the `Log` model itself.
2. no longer store log data in `jsonb` column that contains a bunch of
   duplicate information for each log entry; instead create separate new
   tables, one which will store numeric values and one which will store
   text values

There is some remaining cleanup that can still be done (for example,
deleting the `LogInput` model), but some of it needs to wait until after
this change deploys (e.g. so that the `LogInput` model can facilitate
one of the datamigrations in this PR).