Use $(...) rather than backticks in rubocop command in .travis.yml
This suggestion was made by ShellCheck. ShellCheck also suggested
quoting the value (https://github.com/koalaman/shellcheck/wiki/SC2046),
but we don't want to do that, because we actually do want the output of
git ls-tree -r HEAD --name-only to be handled as separate, space
separated "words".

Also, remove unnecessary use of quotation marks in the YAML for our
commands to verify versions.