Move groceries#show action to groceries#index
"Index" seems like the fundamental/base/root route for a resource, and
since the resource in question is the groceries app, it seems
appropriate to serve it via "index". "Show" would be more appropriate if
we were showing some subset of the "groceries resource" or the
"groceries app", but we aren't -- we're showing the whole groceries app.