Use router-link in quick selector rather than manual <a> tags
The motivation for this change is that it allows CMD-clicking a link in
the quick selector to open the link in a new tab; this wasn't working
when using `<a>` tags without `href`s. (An alternative solution would
have probably been to just give an `href`, but it's probably best to let
the `router-link` component do that, rather than doing so manually.)

Also, slightly related change: remove "All logs" option from the quick
selector. It seemed like a sort of good idea, at the time, but there's
not really much of a point. When viewing the quick-selector, one is
_already_ viewing a list of all logs (assuming that no search string has
been entered).

I'm also removing some arguably overly cautious code. Maybe by doing so
we will get some Rollbar errors if these unexpected situations occur.