Update @babel/preset-env dependencies
`yarn upgrade @babel/preset-env`

I ran this command because I closed PR #1666 while dependabot said that
it was doing rebasing of that PR. So I am trying to make sure that all
is in order with the dependencies of `@babel/preset-env`.

For an unknown reason, this added an `integrity` attribute for many
dependencies listed in `yarn.lock`. I am guessing that this is because
dependabot uses a different `yarn` version or something (which doesn't
add `integrity` attributes to `yarn.lock`)?