Address CVE-2015-9284
See https://github.com/omniauth/omniauth/pull/ 809 for more info.

Unfortunately, because of the change of "Sign in with Google" from a
link to a button (so that it will be a POST rather than a GET), the
sign-in-with-Google page now looks uglier than ever. That's okay,
though. Making that page look pretty is on my to-do list.