Bump chromedriver from 75.0.3770.90 to 76.0.3809.68
Travis is currently installing Chrome 76; let's update chromedriver to
the same version.

It would be nice to automate this so that I don't have to keep updating
the chromedriver version manually. I think that there is a gem that does
this, although I am hesitant to depend on a gem. Another alternative
would be to make some requests to some Chrome site(s) and parse the
respose(s) to get the appropriate download link programmatically.