Export default Vuex store from groceries/store.js
This will make it possible for multiple different files to
import/reference the groceries store. (I was motivated to do this
because I was going to use the groceries Vuex store in
`app/javascript/logs/router.js`, but I ultimately went a different
direction. But this still seemed like a good change to commit.)

We leave the factory function as a named export for use in specs.