Close logs quick-selector when route changes; stop managing manually
This change fixes a bug wherein command-clicking (Mac) a log link from
within the logs quick-selector would close the quick-selector. The quick
selector should remain open so that multiple logs can be opened from the
quick selector via command-clicking.

Since command-clicking doesn't trigger a route change within the current
tab, this change (close the logs quick-selector when the route changes,
and stop managing the appearance of the quick-selector manually) results
in the quick selector remaining open when command-clicking links, as
desired.

At first I was going to move the quick-selector state
(`highlightedLogIndex`, `searchString`) into the Vuex store, so that the
router could dispatch Vuex actions to reset the quick selector state
when the route changes.

However, I ended up instead just having the router emit a
`'groceries:route-changed'` event that can be listened for by the quick
selector, and then the quick selector can manage resetting its own
state. This was a simpler/smaller change, and I think it largely "feels
right" in terms of the router shouldn't necessarily "know" about Vuex
actions to update the state of the logs quick-selector.