Redirect to login_path when blocked by devise's `authenticate` method
When a route is guarded by devise's `authenticate` method and a user is
not logged in, devise redirects to `new_user_session_url`, if it exists.
However, we have no such route (instead, we have `login_url`). This
commit adds a `CustomFailureApp` (inheriting from `Devise::FailureApp`)
that will redirect to that path (`login_path`) if a user tries to access
an `authenticate` d route while not logged in.