Change REQUEST_DATA_TTL from 1.hour to 21.days; delete data after use
Since Sidekiq will retry jobs for up to ~21 days, we want to leave the
data in Redis for up to that long.

However, in order to avoid having a bunch of no-longer-needed data build
up in Redis because of this new, longer TTL, we will now also explicitly
delete the data from Redis after we successfully create a request from
it, rather than just waiting for the TTL to expire (as we were doing
before).