Reduce the number of threads, processes, & Redis connections used
This sets Sidekiq concurrency to 2, which is a significant decrease from
the default (which I think is 10 and used to be 25). That's fine; we
aren't really in a big rush to process our Sidekiq jobs. They can wait
for each other. :)

This changes the fallback `WEB_CONCURRENCY` to 2, reflecting a change I
have already made to the Heroku environment variable setting.