Remove FastFeedback project from homepage
I am sick of paying for the fastfeedback.io and ffb.pw domain names, and
I'm also sick of performing security updates for this app, so I am going
to stop hosting a live version of this project. I _could_ still leave up
the project description and GitHub link, but that's kind of lame without
a live version to back it up, plus it uses Backbone.js which isn't
really valuable these days at all, anyway.