Remove "App Academy Picture Flashcard App" project
I just deleted the `aa-picture-game` repo, because it had PII of App
Academy students (names and photos), which they might not want to be on
the public record associated with this game etc. Therefore, the live
site and the GitHub links don't work anymore. Therefore, let's remove
this project from the davidrunger.com homepage.