Simplify Api::LogEntriesController#index action
Previously, we were returning differently formatted JSON data, depending
on whether the request was for the log entries of a single log or the
log entries of all of the current user's logs.

This change simplifies the controller action by always returning the log
entries data in the same format, regardless of whether we are returning
the logs entries of one log or many logs.

This basically just moves the complexity from the backend to the
frontend... but that feels slightly better. This complexity sort of
feels like something that the frontend should manage.