Add `thread_safe` gem (required by active_model_serializers)
`active_model_serializers` depends on this gem (`thread_safe`) but
doesn't declare it as a dependency.

`rails` used to have it as a dependency, but that was recently
[removed][1].

[1]: https://github.com/rails/rails/commit/bf34c808f9936ecb160914cace45c655ec0924c7#diff-e79a60dc6b85309ae70a6ea8261eaf95

Therefore, as a workaround, we'll add `thread_safe` to our own Gemfile.

We can undo this change if/when `active_model_serializers` resolves this
bug (i.e. when they explicitly add `thread_safe` as a dependency or they
remove the dependency on `thread_safe`).