Don't "commit" change to grocery item name until input is blurred
This avoids a UX bug wherein changing a grocery item's name could change
its sort order in the list of grocery items, causing the input to move
up or down in the list while editing. With this change, the item will
only be resorted after blurring the input (since that is when we'll
actually update `item.name`).

To accomplish this, we are now using one-way binding rather than two-way
binding via `v-model`.