Use #include? rather than in? in sidekiq.yml
Evidently (based on H10 errors in production when #2744 / 1b86711
deployed), the `Object#in?` monkeypatch provided by ActiveSupport hasn't
(necessarily?) yet been applied when Sidekiq needs it to have been while
parsing the config file.

It's not clear to me why this error didn't occur when I booted Sidekiq
locally but I guess that it has to do with differences in the boot
process (e.g. `spring`, maybe?).