Log full request path, including query string
For whatever reason, `lograge` [removes][1] the query string from the
logged data. I prefer to see the query string as part of the path. (The
query string params do show up in the `params` that are all logged, but
I we don't know, for example, if a param came from a query string or
from a POST request body, which might be worth knowing.)

[1]: https://github.com/roidrage/lograge/blob/1729eab/lib/lograge/log_subscribers/action_controller.rb#L31