Allow users to create log entry via query param in logs#index action
In order for this to be somewhat secure / prevent CSRF-like attacks, we
also need the user to incude an authorization token as a query param, as
well.

For example, a user might make a request like this:
```
GET /logs/weight?new_entry=120.4&auth_token=123ThisShouldBeTheUsersToken
```

and that will create a new log entry with the value `120.4` for the
user's `weight` log.

The motivation for this feature is that one can then set up a
bookmark/"search" in their browser like this (note the `%s` part):
```
https://www.davidrunger.com/logs/weight?new_entry=%s&auth_token=abc123
```

and set that up with a keyword of `w`, which makes submitting a log
entry for that log as simple as entering `w 120.4` or whatever in the
user's browser URL bar.

(In order to set up that sort of browser bookmark/"search", the user
will need to get their `auth_token`, which, as of this PR, they can do
via the user edit page.)