Use `AWS_*` rather than `S3_*` env vars for Active Storage config
Currently, in production, we have both `S3_*` and `AWS_*` env vars set
to the same values. We need the `AWS_` env vars so that they can be
picked up by the `aws-sdk-s3` gem (used in the `RunHeat` worker). Let's
also use those env vars in our Active Storage config, so we can delete
the `S3_*` env vars and thereby eliminate that duplication.