Have a participant join the new quiz in feature spec
Part of the motivation for this change is to ensure that the
websocket/cable connection to QuizzesChannel completes before the spec
ends. Prior to this change, it wouldn't always do so, causing
`QuizzesChannel#subscribed` to appear to fall in and out of being
covered by tests (as monitored by Codecov).